- reformatted code - moved JsonUtils and XmlUtils to 'helper' package since these classes are also helpful for clients
